Made large improvements to prefabs and projectile shooting as well as added the basics of an AOE script

+++ 3D Game Project/Assets/Scripts/DamageDealer.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DamageDealer : MonoBehaviour
 {
-    public int damage = 5;
+    public float damage = 5;
 
     private void OnCollisionEnter(Collision collision)
     {
